+++ src/detail_page_handle.js
@@ -1,7 +1,7 @@
 const Apify = require('apify'); // eslint-disable-line no-unused-vars
 const Puppeteer = require('puppeteer'); // eslint-disable-line
 
-const typedefs = require('./typedefs'); // eslint-disable-line no-unused-vars
+const { ScrapingOptions, AddressParsed, PlaceUserData } = require('./typedefs'); // eslint-disable-line no-unused-vars
 const ErrorSnapshotter = require('./error-snapshotter'); // eslint-disable-line no-unused-vars
 const Stats = require('./stats'); // eslint-disable-line no-unused-vars
 
@@ -19,7 +19,7 @@ const { log } = Apify.utils;
 *  request: Apify.Request,
 *  searchString: string,
 *  session: Apify.Session,
-*  scrapingOptions: typedefs.ScrapingOptions,
+*  scrapingOptions: ScrapingOptions,
 *  errorSnapshotter: ErrorSnapshotter,
 *  stats: Stats,
 * }} options
@@ -30,7 +30,7 @@ module.exports.handlePlaceDetail = async (options) => {
     } = options;
     const {
         includeHistogram, includeOpeningHours, includePeopleAlsoSearch,
-        maxReviews, maxImages, additionalInfo, geo, placesCache, reviewsSort,
+        maxReviews, maxImages, additionalInfo, geo, placesCache, reviewsSort, reviewsTranslation,
     } = scrapingOptions;
     // Extract basic information
     await waitForGoogleMapLoader(page);
@@ -42,42 +42,42 @@ module.exports.handlePlaceDetail = async (options) => {
         throw 'The page didn\'t load fast enough, this will be retried';
     }
 
-    const pageData = await extractPageData({ page });
+    // Add info from listing page
+    const { rank, searchPageUrl, addressParsed, isAdvertisement } = /** @type {PlaceUserData} */ (request.userData);
+
+    // Adding addressParsed there so it is nicely together in JSON
+    const pageData = await extractPageData({ page, addressParsed });
 
     // Extract gps from URL
     // We need to URL will be change, it happened asynchronously
     await page.waitForFunction(() => window.location.href.includes('/place/'));
     const url = page.url();
 
-    const locationMatch = url.match(/!3d([0-9\-.]+)!4d([0-9\-.]+)/);
-    const latMatch = locationMatch ? locationMatch[1] : null;
-    const lngMatch = locationMatch ? locationMatch[2] : null;
+    const coordinatesMatch = url.match(/!3d([0-9\-.]+)!4d([0-9\-.]+)/);
+    const latMatch = coordinatesMatch ? coordinatesMatch[1] : null;
+    const lngMatch = coordinatesMatch ? coordinatesMatch[2] : null;
 
-    const location = latMatch && lngMatch ? { lat: parseFloat(latMatch), lng: parseFloat(lngMatch) } : null;
-
-    // Add info from listing page
-    const { shownAsAd, rank, searchPageUrl } =
-        /** @type {{shownAsAd:boolean, rank:number, searchPageUrl: string}} */ (request.userData);
+    const coordinates = latMatch && lngMatch ? { lat: parseFloat(latMatch), lng: parseFloat(lngMatch) } : null;
 
     placesCache.addLocation(request.uniqueKey, location, searchString);
     // check if place is inside of polygon, if not return null, geo non-null only for country/state/city/postal
-    if (geo && location && !checkInPolygon(geo, location)) {
+    if (geo && coordinates && !checkInPolygon(geo, coordinates)) {
         // cache place location to keyVal store
         log.warning(`[PLACE]: Place is outside of required location (polygon), skipping... url --- ${url}`);
         stats.outOfPolygon();
-        stats.addOutOfPolygonPlace({ url, searchPageUrl, location });
+        stats.addOutOfPolygonPlace({ url, searchPageUrl, coordinates });
         return;
     }
 
     const detail = {
         ...pageData,
-        shownAsAd,
+        isAdvertisement,
         rank,
         placeId: request.uniqueKey,
         url,
         searchPageUrl,
         searchString,
-        location,
+        location: coordinates, // keeping backwards compatible even though coordinates is better name
         scrapedAt: new Date().toISOString(),
         ...includeHistogram ? await extractPopularTimes({ page }) : {},
         openingHours: includeOpeningHours ? await extractOpeningHours({ page }) : undefined,
@@ -85,7 +85,7 @@ module.exports.handlePlaceDetail = async (options) => {
         additionalInfo: additionalInfo ? await extractAdditionalInfo({ page }) : undefined,
         ...await errorSnapshotter.tryWithSnapshot(
             page,
-            async () => extractReviews({ page, totalScore: pageData.totalScore, maxReviews, reviewsSort }),
+            async () => extractReviews({ page, totalScore: pageData.totalScore, maxReviews, reviewsSort, reviewsTranslation }),
             { name: 'Reviews extraction' },
         ),
         imageUrls: await errorSnapshotter.tryWithSnapshot(